Fixed Weapon Class Hierarchy, added Grenade
Now Weapon.cs is the parent of all weapons and has virtual functions
Shoot()
Release()
Effect()

These can be overridden by the children classes. They have default behaviors.

Now, in Player.cs, the player listens for buttonUp because we need to time releasing the weapons.
There is a release function that makes sure the player is holding a weapon before we call release
on the weapon.

Weapon Spawn now spawns grenades as well.

Theres a script called Expire.cs that automatically kills the given game object after a certain period
of time.