Implemented Grappling Hook and Weapon Manager
The weapon spawn points no longer spawn weapons, only the weapon manager does it.
The weapon manager spawns all weapons and consumable (except baha blast because it was broken for some reason)
the grappling hook pulls the other character toward the player who used it and puts the player who used it in front of
the affected player.